Refactor Error and move `init`
