+++ lazy-pages/src/sys/windows.rs
@@ -1,3 +1,23 @@
+/*
+ * This file is part of Gear.
+ *
+ * Copyright (C) 2022 Gear Technologies Inc.
+ * SPDX-License-Identifier: GPL-3.0-or-later WITH Classpath-exception-2.0
+ *
+ * This program is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
+ * GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License
+ * along with this program. If not, see <https://www.gnu.org/licenses/>.
+ */
+
 use crate::sys::ExceptionInfo;
 use std::io;
 use winapi::{
@@ -28,12 +48,12 @@ unsafe extern "system" fn exception_handler(exception_info: *mut EXCEPTION_POINT
         fault_addr: addr as *mut _,
     };
 
-    super::memory_exception_handler(info).expect("Memory exception handler");
+    super::user_signal_handler(info).expect("Memory exception handler");
 
     EXCEPTION_CONTINUE_EXECUTION
 }
 
-pub unsafe fn setup_memory_exception_handler() -> io::Result<()> {
+pub unsafe fn setup_signal_handler() -> io::Result<()> {
     SetUnhandledExceptionFilter(Some(exception_handler));
     Ok(())
 }
